+++ CHANGELOG.md
@@ -32,6 +32,8 @@ changes if the major version hasn't changed.
   messages that can be sent to studio. (#28)
 - `fiberplane-models`: Added `ArrayField` variant of `QueryField` to specify provider
   query schemas that include rows/records of arbitrary data. (#39)
+- `fiberplane-models`: Added various structures for webhooks. (#41)
+- `fiberplane-models`: Added new webhook endpoints. (#41)
 
 ### Changed